backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
pinecone-client==4.1.0
openai==1.35.7
sse-starlette==2.1.0
httpx==0.27.0
